Fix up win64 warnings about NtQuerySystemInformation.
svn path=/branches/ros-amd64-bringup/; revision=36274

+++ reactos/base/setup/usetup/interface/usetup.c
@@ -699,7 +699,7 @@ SetupStartPage(PINPUT_RECORD Ir)
     INFCONTEXT Context;
     PWCHAR Value;
     UINT ErrorLine;
-    SIZE_T ReturnSize;
+    ULONG ReturnSize;
 
     CONSOLE_SetStatusText(MUIGetString(STRING_PLEASEWAIT));
